alerts_engine/requirements.txt
python-dotenv
requests
feedparser
google-genai
Pillow
trafilatura
pytrends
